requirements.txt
asgiref==3.5.2
bleach==5.0.0
Django==4.0.5
django-crispy-forms==1.14.0
django-summernote==0.8.20.0
django-taggit==3.0.0
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
whitenoise==6.2.0
